Only include Android Annotations API dependency to build
Both dependencies contain the same set of annotations, hence when trying to
build the project we get an already included error as the same classes are
added twice.  Leaving out the implementation means that annotation processing
does not happen with the build process, and thus is limited to the support
of IDEs such as IDEA.

In the future I imagine additional Gradle configuration to do annotation
processing in the command line, as well as configure IDEs automatically as
configuring them for annotation processing is currently done by hand.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>